requirements.txt
telethon
sqlalchemy
psycopg2-binary
python-dotenv
thefuzz
python-Levenshtein